+++ src/cell.rs
@@ -5,9 +5,9 @@
 #[derive(Clone, Copy)]
 pub struct Cell<const C: usize> {
     /// the current entropy of the cell, or the amount of still possible tiles
-    pub entropy: usize,
+    pub entropy: u8,
     /// the indices of all tiles which are currently possible in this cell; only the entries until the self.entropy index are used
-    tile_indices: [usize; C],
+    tile_indices: [u8; C],
 }
 
 impl<const C: usize> Cell<C> {
@@ -15,29 +15,29 @@ impl<const C: usize> Cell<C> {
         let mut tile_indices = [0; C];
 
         for i in 0..num_tiles {
-            tile_indices[i] = i;
+            tile_indices[i] = i as u8;
         }
 
         Cell {
-            entropy: num_tiles,
+            entropy: num_tiles as u8,
             tile_indices,
         }
     }
 
-    pub fn get_possible_indices(&self) -> &[usize] {
-        &self.tile_indices[0..self.entropy]
+    pub fn get_possible_indices(&self) -> &[u8] {
+        &self.tile_indices[0..self.entropy as usize]
     }
 
-    pub fn collapse(&mut self, index: usize) {
+    pub fn collapse(&mut self, index: u8) {
         self.tile_indices[0] = index;
         self.entropy = 1;
     }
 
-    pub fn set_indices(&mut self, indices: impl IntoIterator<Item=usize>) {
+    pub fn set_indices(&mut self, indices: impl IntoIterator<Item=u8>) {
         let mut entropy = 0;
 
         for (i, index) in indices.into_iter().enumerate() {
-            if index == usize::MAX {
+            if index == u8::MAX {
                 // usize::MAX is used as a placeholder. The first occurrence
                 // of it tells that there are no more indices to consider
                 break;
@@ -50,7 +50,7 @@ impl<const C: usize> Cell<C> {
         self.entropy = entropy
     }
 
-    pub fn get_collapsed_index(&self) -> usize {
+    pub fn get_collapsed_index(&self) -> u8 {
         self.tile_indices[0]
     }
